+++ cli/src/test/kotlin/org/islandoftex/arara/cli/utils/ExecutionTest.kt
@@ -101,13 +101,13 @@ class ExecutionTest : ShouldSpec({
         val output = outputForTest("halt")
         output shouldContain "QuackOne"
         output shouldNotContain "QuackTwo"
-        Arara.exitCode shouldBe 0
+        Executor.executionStatus.exitCode shouldBe 0
     }
     should("forcefully halt on halt error rule") {
         val output = outputForTest("halt-error")
         output shouldContain "QuackOne"
         output shouldNotContain "QuackTwo"
-        Arara.exitCode shouldNotBe 0
+        Executor.executionStatus.exitCode shouldNotBe 0
     }
 
     should("fail on invalid config") {