+++ core/src/main/kotlin/org/islandoftex/arara/core/session/Executor.kt
@@ -9,6 +9,7 @@ import org.islandoftex.arara.api.files.Project
 import org.islandoftex.arara.api.files.ProjectFile
 import org.islandoftex.arara.api.rules.Directive
 import org.islandoftex.arara.api.session.ExecutionReport
+import org.islandoftex.arara.api.session.ExecutionStatus
 import org.islandoftex.arara.api.session.Executor
 import org.islandoftex.arara.core.dependencies.ProjectGraph
 import org.islandoftex.arara.core.files.byPriority
@@ -40,6 +41,11 @@ object Executor : Executor {
      * Specify custom user hooks to run.
      */
     var hooks = ExecutorHooks()
+
+    /**
+     * The execution status this executor is captured in.
+     */
+    var executionStatus: ExecutionStatus = ExecutionStatus.PROCESSING
 
     /**
      * The setup for all executions run by this executor. The execution options